+++ backend/windmill-api/src/lib.rs
@@ -103,6 +103,7 @@ mod inkeep_ee;
 mod inkeep_oss;
 mod inputs;
 mod integration;
+mod live_migrations;
 #[cfg(feature = "postgres_trigger")]
 mod postgres_triggers;
 
@@ -623,7 +624,7 @@ pub async fn run_server(
                         .nest("/mqtt_triggers", mqtt_triggers_service)
                         .nest("/sqs_triggers", sqs_triggers_service)
                         .nest("/gcp_triggers", gcp_triggers_service)
-                    .nest("/postgres_triggers", postgres_triggers_service),
+                        .nest("/postgres_triggers", postgres_triggers_service),
                 )
                 .nest("/workspaces", workspaces::global_service())
                 .nest(
