+++ backend/windmill-worker/src/dedicated_worker.rs
@@ -622,8 +622,7 @@ async fn spawn_dedicated_worker(
                 } else {
                     sqlx::query_as::<_, (String, Option<String>, Option<ScriptLang>, Option<Vec<String>>, bool, Option<ScriptHash>)>(
                         "SELECT content, lock, language, envs, codebase IS NOT NULL, hash FROM script WHERE path = $1 AND workspace_id = $2 AND
-                            created_at = (SELECT max(created_at) FROM script WHERE path = $1 AND workspace_id = $2 AND
-                            deleted = false AND lock IS not NULL AND lock_error_logs IS NULL)",
+                            archived = false AND lock IS not NULL AND lock_error_logs IS NULL ORDER BY created_at DESC LIMIT 1",
                     )
                     .bind(&path)
                     .bind(&w_id)
